Started working on the edit form for courses.
There are several complications arising now:

1. The form action: We don't have any URL helpers.
2. Validation: Haven't settled on a validation library.
3. Redisplay: Don't have any form helpers.

Questions is if redisplay is a very hard problem, or something which can
be solved quite simply using basic jade helpers -- simply building an
input as need. Full control over HTML output of forms is nice.

Validation: node-validation or node-forms look nice (or if it was
express-forms -- whatever, the one that you get when googling this
stuff).

We need some kind of URL helpers. The ones that exist seem very
complicated -- I don't know if "named" routes really are better for an
application of this scale than simply using the URL patterns themselves.
Cool URIs don't change! And it's a natural way of referring to paths
anyway (by actually using paths!).