Drops express-resource.
express-resource has a number of oddities and problems that make it hard
to use:

1) using "thing.<format>" in the URLs make it hard to have e.g. e-mail
addresses as IDs (because ".se" is a "format" in "student.hhs.se")
2) it's hard to pass middleware the express way to individual requests.
this makes it hard to use with express-form.

For now, it's simple enough to list routes manually, and rewrite the
.load-function to work as middleware (as express intended). It might get
tedious, eventually. But this is a rather small project.