Very basic U operation on courses.
This is where an ORM is starting to look nicer and nicer. I'm still
hesitant, though. In principle, I think some of the more common patterns
could simply be abstracted away -- but I'm not entirely sure how.

Also, it might start becoming time to think about what should be the
Model in our not-very MVC app. The controllers are growing fat, with
code that's arguably boilerplate. I'm interested in having PostgreSQL be
the model as much as possible -- speaking with it over SQL. This goes
against a lot of the recent development in web development, which treats
the database as "stupid" and puts as much as possible in the
"application". Well... okay. But maybe data is important and should be
handled by a database, and the application is really mostly an
interface.